+++ src/navigation/action/preselectCharacterTwoStyle.action.js
@@ -1,0 +1,8 @@
+export const PRESELECT_CHARACTER_TWO_STYLE = "preselectCharacterTwoStyle";
+
+export default function preselectCharacterTwoStyle(styleIndex) {
+  return {
+    type: PRESELECT_CHARACTER_TWO_STYLE,
+    styleIndex
+  };
+}