+++ src/navigation/navigation.reducer.js
@@ -1,6 +1,7 @@
 import { PRESELECT_CHARACTER_ONE } from "./action/preselectCharacterOne.action";
 import { SELECT_CHARACTER_ONE } from "./action/selectCharacterOne.action";
 import { SELECT_CHARACTER_ONE_WITH_STYLE_AND_COLOR } from "./action/selectCharacterOneWithSyleAndColor.action";
+import { PRESELECT_CHARACTER_ONE_STYLE } from "./action/preselectCharacterOneStyle.action";
 import { SELECT_CHARACTER_ONE_STYLE } from "./action/selectCharacterOneStyle.action";
 import { PRESELECT_CHARACTER_ONE_COLOR } from "./action/preselectCharacterOneColor.action";
 import { SELECT_CHARACTER_ONE_COLOR } from "./action/selectCharacterOneColor.action";
@@ -8,6 +9,7 @@ import { UNSELECT_CHARACTER_ONE } from "./action/unselectCharacterOne.action";
 import { PRESELECT_CHARACTER_TWO } from "./action/preselectCharacterTwo.action";
 import { SELECT_CHARACTER_TWO } from "./action/selectCharacterTwo.action";
 import { SELECT_CHARACTER_TWO_WITH_STYLE_AND_COLOR } from "./action/selectCharacterTwoWithSyleAndColor.action";
+import { PRESELECT_CHARACTER_TWO_STYLE } from "./action/preselectCharacterTwoStyle.action";
 import { SELECT_CHARACTER_TWO_STYLE } from "./action/selectCharacterTwoStyle.action";
 import { PRESELECT_CHARACTER_TWO_COLOR } from "./action/preselectCharacterTwoColor.action";
 import { SELECT_CHARACTER_TWO_COLOR } from "./action/selectCharacterTwoColor.action";
@@ -19,6 +21,7 @@ import { END_FIGHT } from "./action/endFight.action";
 import preselectCharacterOne from "./transition/preselectCharacterOne.transition";
 import selectCharacterOne from "./transition/selectCharacterOne.transition";
 import selectCharacterOneWithStyleAndColor from "./transition/selectCharacterOneWithStyleAndColor.transition";
+import preselectCharacterOneStyle from "./transition/preselectCharacterOneStyle.transition";
 import selectCharacterOneStyle from "./transition/selectCharacterOneStyle.transition";
 import preselectCharacterOneColor from "./transition/preselectCharacterOneColor.transition";
 import selectCharacterOneColor from "./transition/selectCharacterOneColor.transition";
@@ -26,6 +29,7 @@ import unselectCharacterOne from "./transition/unselectCharacterOne.transition";
 import preselectCharacterTwo from "./transition/preselectCharacterTwo.transition";
 import selectCharacterTwo from "./transition/selectCharacterTwo.transition";
 import selectCharacterTwoWithStyleAndColor from "./transition/selectCharacterTwoWithStyleAndColor.transition";
+import preselectCharacterTwoStyle from "./transition/preselectCharacterTwoStyle.transition";
 import selectCharacterTwoStyle from "./transition/selectCharacterTwoStyle.transition";
 import preselectCharacterTwoColor from "./transition/preselectCharacterTwoColor.transition";
 import selectCharacterTwoColor from "./transition/selectCharacterTwoColor.transition";
@@ -49,6 +53,9 @@ export default function navigationReducer(data, action) {
     case SELECT_CHARACTER_ONE_WITH_STYLE_AND_COLOR:
       return selectCharacterOneWithStyleAndColor(data, action);
 
+    case PRESELECT_CHARACTER_ONE_STYLE:
+      return preselectCharacterOneStyle(data, action);
+
     case SELECT_CHARACTER_ONE_STYLE:
       return selectCharacterOneStyle(data, action);
 
@@ -69,6 +76,9 @@ export default function navigationReducer(data, action) {
 
     case SELECT_CHARACTER_TWO_WITH_STYLE_AND_COLOR:
       return selectCharacterTwoWithStyleAndColor(data, action);
+
+    case PRESELECT_CHARACTER_TWO_STYLE:
+      return preselectCharacterTwoStyle(data, action);
 
     case SELECT_CHARACTER_TWO_STYLE:
       return selectCharacterTwoStyle(data, action);